requirements.txt
fastapi
uvicorn
supabase

python-dateutil
pytz
jinja2
python-multipart
requests
python-dotenv
